src/local_analysis/src/requirements.txt
av==10.0.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.4.0
et-xmlfile==1.1.0
fonttools==4.39.3
fsspec==2023.4.0
importlib-metadata==6.6.0
kiwisolver==1.4.4
locket==1.0.0
matplotlib==3.7.1
nd2==0.5.3
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
partd==1.4.0
Pillow==9.5.0
polars==0.20.10
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
resource-backed-dask-array==0.1.0
six==1.16.0
toml==0.10.2
toolz==0.12.0
typing_extensions==4.5.0
zipp==3.15.0
pyinstaller==5.5
